refactor: update market prices for items and adjust recipe production amounts

+++ apps/gersang_db/priv/repo/migrations/20240226160600_create_gersang_items.exs
@@ -5,8 +5,8 @@ defmodule GersangDb.Repo.Migrations.CreateGersangItems do
     create table(:gersang_items) do
       add :name, :string, null: false
       add :tags, {:array, :string}
-      add :margin,  :float
-      add :market_price,  :integer
+      add :margin, :float
+      add :market_price,  :bigint
       add :cost_per, :float
       add :artisan_product?, :boolean, null: false, default: false
       add :artisan_production_amount, :integer
